transphorm getIsSearchPreviousIew into getIsVEnuePreviousView
